cloud_functions/api-categories/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-firestore
